backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0
asyncpg>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pandas>=2.2.0
ezdxf>=1.1.0
python-magic>=0.4.27
alembic>=1.13.0
psycopg2-binary>=2.9.9